# backend/src/api/generated/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
    baseUrl: `${string}://${string}` | (string & {});
};

/**
 * A guest booking. The end time is calculated from the event type duration.
 */
export type Booking = {
    id: Uuid;
    eventTypeId: Uuid;
    eventTypeTitle: string;
    startsAt: string;
    /**
     * Calculated as startsAt plus the selected event type duration.
     */
    endsAt: string;
    guestName: string;
    guestEmail: string;
};

/**
 * Input used by a guest to create a booking.
 */
export type CreateBookingRequest = {
    /**
     * The event type selected by the guest.
     */
    eventTypeId: Uuid;
    /**
     * Must be one of the free slot start times for the selected event type.
     */
    startsAt: string;
    guestName: string;
    guestEmail: string;
};

/**
 * Input used by the calendar owner to create an event type.
 */
export type CreateEventTypeRequest = {
    title: string;
    description: string;
    /**
     * Duration of this event type in minutes.
     */
    durationMinutes: number;
};

export type ErrorResponse = {
    code: string;
    message: string;
};

/**
 * A type of event that guests can book.
 */
export type EventType = {
    id: Uuid;
    title: string;
    description: string;
    /**
     * Duration of this event type in minutes.
     */
    durationMinutes: number;
};

/**
 * A free booking slot for a selected event type in a calendar month.
 */
export type Slot = {
    startsAt: string;
    endsAt: string;
};

/**
 * Database-generated UUID.
 */
export type Uuid = string;

/**
 * Calendar month in YYYY-MM format.
 */
export type YearMonth = string;

export type AdminBookingsListUpcomingData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/admin/bookings/upcoming';
};

export type AdminBookingsListUpcomingResponses = {
    /**
     * The request has succeeded.
     */
    200: Array<Booking>;
};

export type AdminBookingsListUpcomingResponse = AdminBookingsListUpcomingResponses[keyof AdminBookingsListUpcomingResponses];

export type AdminEventTypesCreateData = {
    body: CreateEventTypeRequest;
    path?: never;
    query?: never;
    url: '/admin/event-types';
};

export type AdminEventTypesCreateResponses = {
    /**
     * The request has succeeded and a new resource has been created as a result.
     */
    201: EventType;
};

export type AdminEventTypesCreateResponse = AdminEventTypesCreateResponses[keyof AdminEventTypesCreateResponses];

export type BookingsCreateData = {
    body: CreateBookingRequest;
    path?: never;
    query?: never;
    url: '/bookings';
};

export type BookingsCreateErrors = {
    /**
     * The server could not understand the request due to invalid syntax.
     */
    400: ErrorResponse;
    /**
     * The server cannot find the requested resource.
     */
    404: ErrorResponse;
    /**
     * The request conflicts with the current state of the server.
     */
    409: ErrorResponse;
};

export type BookingsCreateError = BookingsCreateErrors[keyof BookingsCreateErrors];

export type BookingsCreateResponses = {
    /**
     * The request has succeeded and a new resource has been created as a result.
     */
    201: Booking;
};

export type BookingsCreateResponse = BookingsCreateResponses[keyof BookingsCreateResponses];

export type EventTypesListData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/event-types';
};

export type EventTypesListResponses = {
    /**
     * The request has succeeded.
     */
    200: Array<EventType>;
};

export type EventTypesListResponse = EventTypesListResponses[keyof EventTypesListResponses];

export type EventTypesReadData = {
    body?: never;
    path: {
        eventTypeId: Uuid;
    };
    query?: never;
    url: '/event-types/{eventTypeId}';
};

export type EventTypesReadErrors = {
    /**
     * The server cannot find the requested resource.
     */
    404: ErrorResponse;
};

export type EventTypesReadError = EventTypesReadErrors[keyof EventTypesReadErrors];

export type EventTypesReadResponses = {
    /**
     * The request has succeeded.
     */
    200: EventType;
};

export type EventTypesReadResponse = EventTypesReadResponses[keyof EventTypesReadResponses];

export type EventTypeSlotsListData = {
    body?: never;
    path: {
        eventTypeId: Uuid;
    };
    query?: {
        month?: YearMonth;
        timeZone?: string;
    };
    url: '/event-types/{eventTypeId}/slots';
};

export type EventTypeSlotsListErrors = {
    /**
     * The server could not understand the request due to invalid syntax.
     */
    400: ErrorResponse;
    /**
     * The server cannot find the requested resource.
     */
    404: ErrorResponse;
};

export type EventTypeSlotsListError = EventTypeSlotsListErrors[keyof EventTypeSlotsListErrors];

export type EventTypeSlotsListResponses = {
    /**
     * The request has succeeded.
     */
    200: Array<Slot>;
};

export type EventTypeSlotsListResponse = EventTypeSlotsListResponses[keyof EventTypeSlotsListResponses];